+++ source/extensions/omni.flux.validator.mass.widget/omni/flux/validator/mass/widget/setup_ui.py
@@ -30,6 +30,8 @@ from omni.flux.utils.common import EventSubscription as _EventSubscription
 from omni.flux.utils.common import reset_default_attrs as _reset_default_attrs
 from omni.flux.validator.mass.core import Executors as _MassExecutors
 from omni.flux.validator.mass.core import ManagerMassCore as _ManagerMassCore
+from omni.flux.validator.mass.core.executors import CurrentProcessExecutor as _CurrentProcessExecutor
+from omni.flux.validator.mass.core.executors import ExternalProcessExecutor as _ExternalProcessExecutor
 from omni.flux.validator.mass.queue.widget import Actions as _MassQueueTreeActions
 from omni.flux.validator.mass.queue.widget import MassQueueTreeWidget as _MassQueueTreeWidget
 from omni.flux.validator.mass.queue.widget.tree.delegate import Delegate as _Delegate
@@ -99,8 +101,11 @@ class ValidatorMassWidget:
             "_schema_tree_view_tab_toggled": None,
             "_mass_queue_frame": None,
             "_executors_cb": None,
+            "_executor_container": None,
             "_queue_tree_model": None,
             "_queue_tree_delegate": None,
+            "_current_process_executor": None,
+            "_external_process_executor": None,
         }
         for attr, value in self._default_attr.items():
             setattr(self, attr, value)
@@ -131,6 +136,9 @@ class ValidatorMassWidget:
 
         self._queue_tree_model = _Model()
         self._queue_tree_delegate = _Delegate(use_global_style=self._use_global_style, style=self._style)
+
+        self._current_process_executor = _CurrentProcessExecutor()
+        self._external_process_executor = _ExternalProcessExecutor()
 
         self.__on_mass_queue_action_pressed = _Event()
         self.__on_schema_tab_toggled = _Event()
@@ -276,6 +284,7 @@ class ValidatorMassWidget:
                     result_run: List[Tuple["_ManagerCore", asyncio.Future]] = await self._core.create_tasks(
                         self._executors_cb.model.get_item_value_model().get_value_as_int(),
                         result,
+                        custom_executors=(self._current_process_executor, self._external_process_executor),
                         standalone=False,
                         queue_id=self._mass_queue_widget.get_queue_id(),
                     )
@@ -397,11 +406,21 @@ class ValidatorMassWidget:
                                     )
                                 with ui.HStack(width=ui.Percent(50), spacing=ui.Pixel(8)):
                                     ui.Spacer()
-                                    ui.Label("Executor:", width=0)
+                                    ui.Label("Execute In: ", width=0)
                                     self._executors_cb = ui.ComboBox(
                                         1,
-                                        *[executor.name.replace("_", " ").capitalize() for executor in _MassExecutors],
+                                        *_MassExecutors.get_names(),
+                                        width=ui.Pixel(150),
+                                        identifier="executors_combo_box",
                                     )
+                                    self._executors_cb.model.add_item_changed_fn(self._update_executor_container)
+
+                                    self._executor_container = ui.HStack(width=0)
+                                    with self._executor_container:
+                                        ui.Spacer()
+                                        self._external_process_executor.create_ui()
+                                        ui.Spacer()
+
                                     with ui.VStack(width=0):
                                         ui.Spacer()
                                         self._normals_type_info_icon = _InfoIconWidget(
@@ -410,10 +429,9 @@ class ValidatorMassWidget:
                                                 "It is not recommended to change this setting unless "
                                                 "you are an expert user and understand its effects.\n\n"
                                                 "Options:\n"
-                                                "- Process executor: will run the ingestion in an external process "
-                                                "(recommended)\n"
-                                                "- Async executor: will run the ingestion asynchronously on the main "
-                                                "thread"
+                                                "- External Process: Run the ingestion in external process(es) "
+                                                "(recommended for running multiple ingestions at once)\n"
+                                                "- Current Process: Run the ingestion asynchronously on the main thread"
                                             )
                                         )
                                         ui.Spacer()
@@ -445,6 +463,15 @@ class ValidatorMassWidget:
                         ui.Spacer(width=ui.Pixel(8))
 
                 self._validator_widget_frame = ui.Frame(visible=False, identifier="ValidationWidgetFrame")
+
+    def _update_executor_container(self, *_):
+        with self._executor_container:
+            # For the Process executor, show the processor count dropdown; Show nothing for the current process executor
+            current_executor = self._executors_cb.model.get_item_value_model().get_value_as_int()
+            if current_executor == 1:  # external executor
+                self._external_process_executor.create_ui()
+            else:  # current executor (async)
+                self._executor_container.clear()
 
     def show(self, value: bool):
         if self._mass_queue_widget: